Moved QgsLayerTreeModelLegendNode (+derived classes) to a new file
