+++ src/core/qgslegendrenderer.cpp
@@ -3,6 +3,7 @@
 #include "qgscomposerlegenditem.h"
 #include "qgslayertree.h"
 #include "qgslayertreemodel.h"
+#include "qgslayertreemodellegendnode.h"
 #include "qgslegendmodel.h"
 #include "qgsmaplayerlegend.h"
 #include "qgsmaplayerregistry.h"
