+++ tests/src/core/testqgslegendrenderer.cpp
@@ -8,6 +8,7 @@
 #include "qgsfontutils.h"
 #include "qgslayertree.h"
 #include "qgslayertreemodel.h"
+#include "qgslayertreemodellegendnode.h"
 #include "qgslegendmodel.h"
 #include "qgsmaplayerlegend.h"
 #include "qgsmaplayerregistry.h"
